+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 - add parallel save
 - add shrink-on-load using the "page" param to kakaduload
-- add "options" flag to kakadusave
+- add "options" arg to kakadusave
 - add resolution load and save
 - add "htj2k" arg to enable high-throughput j2k compression
+- add "rate" arg to save
 
 ## 2024/1/11 1.0